Command Execution added, other fixes.
- CommandExe.c executes commands using pathSearch and token arguments
- Path.c had a memory leak caused by not freeing variables in char** array
- Path.c no longer returns static c-string errorMsg causing a seg fault when freed
- Project1.c has commandExecution added to loop
- Project1.c seg fault fixed when user enters an empty string (newline only) as input
- tokenList and its functions defined outside Project1.c to be used elsewhere
- makefile warnings fixed
