+++ Prompt.c
@@ -3,12 +3,12 @@
 //Venkata Sai Pavan Kumar Vadrevu, Daniel Jamsheedy, Luke Power
 
 #include "Prompt.h"
-#include "EnvVariables.h"
 
 void printPrompt(){
   char* user = getEnvironment("$USER");
   char* host = getEnvironment("$HOST");
   char* pwd = getEnvironment("$PWD");
+
   printf("%s@%s: %s>", user, host, pwd);
 
   free(user);
