+++ Prompt.h
@@ -2,11 +2,12 @@
 //COP4610
 //Venkata Sai Pavan Kumar Vadrevu, Daniel Jamsheedy, Luke Power
 
-#ifndef ENV_VARIABLES_H
-#define ENV_VARIABLES_H
+#ifndef PROMPT_H
+#define PROMPT_H
 
 #include <stdio.h>
 #include <stdlib.h>
+#include "EnvVariables.h"
 
 void printPrompt();
 #endif
