+++ makefile
@@ -1,7 +1,7 @@
 all: shell
 
 shell:
-	gcc -o shell Project1.c EnvVariables.c Echo.c Prompt.c Path.c -std=c99
+	gcc -o shell Project1.c tokenlist.c EnvVariables.c Echo.c Prompt.c Path.c CommandExe.c -std=c99
 
 clean:
 	rm shell *.o
